config: move CI() to after config merge
Tests that require a merged config, ex) Test_flag_config_TT, were failing
because the CI() returned prior to merging the data between the global
lab.toml file and the local .git lab.tom file.

Move CI() to after the config merge, and clarify the comment above
the call.  The host, user, and token values will be overwritten by the
environment variables in the CI() function.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>