cmd: use heredoc.Doc() in every multiline help message
Multilines in Golang in general looks pretty bad in the middle of the
code, even more it's presented to the user somehow, since the leading
whitespaces that the developer would usually use as indentation is shown
as actual spaces when the multiline text is printed.

This patch adds the heredoc module and makes use of the heredoc.Doc()
function to get rid of the leading spaces in multiline text, considering
the first line indentation as the start of the actual visual
whitespaces.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>