mr_reopen_test: fix MR number
The MR 19, that was being used as test MR for close/reopen operations,
had as source target a branch from my personal fork that I mistakenly
deleted. With that, the MR cannot be reopened anymore and, therefore,
the tests were failing.

This patch use the new MR created for the same purpose.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>