edit: fix labels update for issue and MR
Since the label/unlabel code was written the go-gitlab module was
updated and received got the ability to add and remove labels explicitly
instead of the specifying all labels through a single struct member.

With that, we can use these new members in the UpdateLabel struct to
get the job done and solve the issue where it's not possible to unlabel
an issue or MR with a single label, since the array would be ignore for
being "empty".

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>