cmd/util: avoid an empty string to be parsed
An additional empty string was being added in the array of "branchArgs" to
later be processed as a valid branch/remote information by the argument
parses. It wasn't an issue before because a "err" return value was being
ignored somewhere else within the parses, but the commit 69ae19e14cd2 ("cmd:
fix return value checks and statements") changed it and the error finally
was cought.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>