mr show: Fix --patch and --comments option output
Users have been setting "comments = true" for the mr show command.  This
causes weird output with the --patch output in which the git log is output,
then the comments from the MR is output.

Make a note that --patch and --comments do not work together and stop
the output after the git log is output.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>