cmd/mr: remote WIP terminology support
GitLab is planning to fully remove the WIP word from their Draft status [1]
in their 14.0 version, which is about to be released. With that, we should
also error out when the user uses it for their merge requests prefixes.

```
Adding WIP: to the start of the merge request’s title still marks a merge
request as a draft.  This feature is scheduled for removal in GitLab 14.0.
Use Draft: instead.
```

[1] https://gitlab.com/gitlab-org/gitlab/-/issues/228685
[2] https://docs.gitlab.com/ee/user/project/merge_requests/drafts.html

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>