internal/git: rename PathWithNameSpace to PathWithNamespace
"Namespace" is a word with a single meaning and is also used within the
go-gitlab module, however, in lab, we have a PathWithNameSpace, with
capiltal 'S' in "NameSpace", which is confusing since has the exact same
meaning as "Namespace" from go-gitlab. With that, this patch renames all
occurences of "NameSpace" to "Namespace" in lab's source code.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>