internal/git: fix git log command with union behavior
In both NumberCommits() and Log() the git command was being used:

	git log <target>...<source>

which is exactly the same as:

	git log <target> <source> --not $(git merge-base --all <target> <source>)

Meaning that a union from both <target> and <branch>, but starting at
the merge base, is returned. However, that's not exactly the behavior we
want: we only want what are the commits that the <source> have that
<target> doesn't (or, B-A, the difference from B to A) which is
essentially the same information presented in the GitLab UI when
creating an MR.

With that, this patch changes both occurrences of 'A...B' to 'A..B',
which has the meaning that we want:

	git log <source> ^<target>

List every commit different in <source> starting from the <target>
point.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>