internal/git: allow specific ref to be used in LastCommitMessage()
LastCommitMessage was querying for the last commit message from HEAD
only, hard-coded. So if the user uses --source option for creating an MR
that doesn't match with the branch he's checked-out the commit log
prompted will be from HEAD instead of what's present in the requested
source remote:branch through --source option.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>