cmd/note_common: use Cobra cmd obj to inspect command name and alias
Instead of checking for os.Args[] directly, use Cobra's command object
to inspect what's the parent command used (issue or mr?) and also to
check if the command `note` was called directly or through its `resolve`
alias.

Even though it's ok to use os.Args, it might be mangled somewhere else,
and using the already existent Cobra infrastructure seems a safer choice
for the job.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>