mr show: Output commit ID hash on commit comments
Output the commid ID hash on commit comments.

Note: The behavior of note.Position.BaseSHA has changed to point to the
commit just before the commented on commit, and note.Position.HeadSHA
has changed to point to the commented on commit.  This gives the ability
to output not only the commid ID hash, but the proper diff.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>