mr show: Show correct commit comments
The first few versions of this code tried to use git to display comments
on diffs using data from the Note.Position field.  A problem was found
with this approach when trying to reference outdated/superseded commits
that were not in the user's tree.

Additionally, the commented on commit ID can now be found in the
Note.CommitID field (this is a recent addition to the GitLab API) and
using the Note.CommitID field a GitLab API call can be made to fetch a
diff for the commit ID.  The Note.Position.LineRange struct contains
information on the exact line that is being commented on.

Show the correct data for both older outdated commits, and current
commits.

Additional change: Add a "|" to the commit output.  The display is easier
to read with this change.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>