rhstatus: Add commit comment functionality
The GitLab WebUI allows users to comment directly on commits.  These comments
appear both in the comment view and the MR view.  This functionality can be
implemented via the lab.Commits.PostCommitComment() function, however, the
function only supports comments on "new" lines of code.  After discussions
with GitLab they have suggested using lab.Discussions.CreateCommitDiscussion()
as a workaround [1].  This solution has a minor issue in that it will not work
with commits that have multiple parents.

Implement the ability for users to comment directly on commits and to
block MRs with commit comments.

[1] https://gitlab.com/gitlab-org/gitlab/-/issues/335337#note_643291819

Signed-off-by: Prarit Bhargava <prarit@redhat.com>