mr_discussion: Provide better discussion information
In edit mode 'lab mr discussion' provides very basic instructions about
comment lines being ignored.  This is useful but it is not enough
information on the discussion thread being created.

Add the status of the MR, the MR ID, and if specified, the commit ID, to
the discussion information in edit mode.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>