mr_create: Add -f option to open editor
A request was made to add an option to open the editor when a file was
specified for the title and description.  Opening the editor would allow
project owners and maintainers to deploy MR templates that could be edited
by the MR submitter.

Add a --file-open_editor/-f option that opens the editor with the file
contents.

Closes #739

Suggested-by: David Arcari <darcari@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>