gitlab: handle API response without X-Total(-Pages) HTTP header
In cases where the number of comments in a certain Issue/MR is greater than
10k the HTTP header for the API responses lack the X-Total and X-Total-Pages
metadata [1] used by lab to control the response pagination. Even though it
should not happen in real scenarios, it's breaking our tests where some MRs
have more than 10k comments. The solution adopted was to use the X-Next-Page
header whenever the X-Total(-Page) is not found.

[1] https://docs.gitlab.com/ee/user/gitlab_com/index.html#pagination-response-headers

Signed-off-by: Lucas Zampieri <lzampier@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>