ci_trace: fix args filter wrt branch name
The CI args handler is considering only the branch name when the user passes
it with the job name in the format `<remote> <branch>:<jobname>` or with an
empty `<jobname>` like `<remote> <branch>:`. In all other subcommands, the
branch name can be used after the remote name without any additional suffix
(the colom in this case ':'). This patch fixes it and adds the ability to
the user call CI subcommands with `<remote> <branch>`.

This patch also adds some comments to the code where this filtering happens,
since the code, even though small, is quite dense wrt the number of
possibilities.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>