Makefile: avoid spurious error messge on first run
My first run of "make test" gave me this harmless error

	make[1]: Entering directory '/home/johannes/git/lab'
	rm coverage-* 2>&1 > /dev/null || true
	rm: cannot remove 'coverage-*': No such file or directory

The redirections are meant to silence the error but fail because
"> /dev/null" does not change the earlier redirection of stderr to
stdout. It would work if written as ">/dev/null 2>&1". Even better,
let's just use "rm -f", then we don't need to check the exit code
either.

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>