mr discussion: add some examples for optional MR ID / branch
The first parameter is the remote, that can be left out because it
defaults to the default remote or origin (see guessDefaultRemote()).
Add an example that uses the default remote; that should be good
enough in many scenarios.

All examples assume that the merge request's source branch is
checked out (because they don't pass an explicit MR id parameter).
Add examples that specify a different MR, either via MR id or by
branch name.

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>