mr: state that the MR ID parameter can be the branch name
The name of a MR's source branch can be used instead of the numeric
MR id parameter on the command line. Advertise this in the command
synopses.

The branch name is easier to remember, and plays better with standard
Git tooling than the opaque MR ids.

While at it, add square brackets to mark that the MR id parameter
is optional.  If it's left out, we use the MR associated with
the branch pointed to by HEAD.

Signed-off-by: Johannes Altmanninger <aclopte@gmail.com>