util: get last pipeline regardless its result
So far, lab was picking the latest pipeline with in good state (passed), but
ignoring any pipeline in case all of them failed. This patch fixes it by
picking the latest pipeline regardless its result, from the `head_pipeline`
API field. With that, we guarantee that the user can fetch information from
failed pipelines and even its artifacts for further investigation.

The plan is to add, in the near future, an option to allow the user to fetch
only pipelines that succeeded.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>