ci: add --passed flag for retrieving only passed pipelines
Today, lab is retrieving information from the latest ran pipeline,
regardless its state. However, there is a use case where the user want to
check the information for only those pipelines that passed.

This patch adds the --passed flag, which does exactly this: retrieve
information for the latest passed pipeline. Also, it adds some testing code
for the new flag and update the mr_list test code to point to the MR created
in the live repository for this flag.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>