root: ignore retval from early cmdline parsing
The first command line parsing happens right at the beginning of the
cmd package creation for setting the debug/quiet flags. This early parsing
results in error when a flag for a subcommand is passed, since flags for the
subcommands were not populated yet. Since unknown flags are going to be
handled later, on subcommand level, we should not care at all by the return
value from the early cmdline parsing. This patch removes this checking,
which was preventing --debug/quiet to be properly set in these cases.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>