gitlab: remove redundant project lookup
The GitLab API request accepts the name of the project passed as normal
string/name (properly scaped) as well as the ID of the project, so it isn't
required for lab to lookup for the project ID when the name of it is known.

This project remove all the FindProject() calls from the internal gitlab
code, decreasing by one the number of API calls made for a single command.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>