config: fix worktree git config dir path
The config code doesn't follow the worktree git config file when the
user calls for a command that depends on that from within a `git
worktree` directory. In this cases, the .git path is in fact a file
instead of a valid directory, which turns to point to the actual
worktree configuration dir.

Instead of realying in a hardcoded path (`.git/lab`), make use of the
`git rev-parse --git-dir` command already available in the git commands
code (internal/git/git.go) and which follows the worktree situation.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>