cmd/root: fix setting log level
Parsing the command line this early in the code won't account for the
subcommands flags, causing the parse to fail. Since later in the code we
already check for the parsing errors considering the subcommands, we can
ignore the parse error for setting the log level in the initialization
code, thus even if some flags are wrong the debug level can be used.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>