cmd/edit: unify description edition functions
Instead of having two different functions for editing MR and issue
descriptions based on the sources (command line or filename), use the
same function but different args. The actions are simple, so keeping
them together is simpler to maintain and understand. In case the
function require additional logic for new things we may split them
again later.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>