internal/gitlab: fix GroupSearch() logic
The group search logic didn't consider the fact that the GitLab API
returns an alphabetical ordered list of groups within a certain group,
thus when multiple subgroups were nested the value groups[0] and list[0]
used in the code weren't always the expected ones.

With that, instead of looking up for the first group, we look for the
last one, meaning that all subgroups found in user's namespace or
wherever he has access to, is returned from the API. The match then
happens with the group full path or full name, which matches first.

In case the user has access to multiple groups with ambiguous names,
all possible paths are shown and lab's exits.

Signed-off-by: Lucas Zampieri <lzampier@redhat.com>
Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>