fork: allow user to fork a project from its numerid ID
The `lab fork` command allows only the project name from user's input, but
it's not a huge change to allow the user to enter the project numeric ID
instead. This commit adds this ability to the fork command by changing the
way Fork() handles the project ID: receiving as interface{} and checking its
real value. With that, an additional change was necessary in the
FindProject(), which was accepting only project IDs as string.

As consiquence, this change allow other functions to pass either the project
ID comming from a go-gitlab module object (int) or the one comming from the
user input (string) to the FindProject() function.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>