gitlab: standardize project ID type
In the internal gitlab module, project IDs were being passed in two
different types in different functions: interface{} and string. This isn't a
problem code-wise because the go-gitlab module accept everything in the
interface{} type and parse the value internally. However, it's confusing for
lab developers to understand when to use which, and honestly, it doesn't
seem to have any good reason.

So, to maintain consistency, change all string project IDs to interface{}
giving more flexibility and also adds the parseID() function to convert
the interface{} ID to string when really necessary.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>