cmd/mr_checkout: check existing remotes before creating one
In case tacking is enabled `lab mr checkout -t`, a new remote is created
with the name of the MR author's name without checking if the already
available remotes matches the one from which the MR was checked-out.
A particular case where it happens constantly is in the branching
development model: developers use a single repo to push their work as
branches.

This patch change the code behavor by first checking if any of the
existent remotes has the same path as the source MR project before
creating a new remote. If no remote is found, then create a new one with
the MR author's name.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>