internal/git: fix remotes lookup
Both calls to Remotes() and RemoteBranches() were limited for looking into the
current directory for the .git folder, meaning that the call wouldn't work
properly if a user was in a repo subdirectory (where .git/ doesn't exist).

One way of solving it would be to check the repo topdir and pass it as the
working directory to the gogit.PlainOpen() call, however, these two functions
are the only ones actually using the go-git module in lab's code. Considerint
it isn't a small dependency to be kept and that every other internal/git/git.go
function relies on the 'git' executable, lets remove the go-git mod dependency
for now and, when really necessary, move all other functions to go-git at once.

Using the 'git' executable defers the topdir lookup to git itself. Thus, we
only need to traverse the list of remotes we have with a simple 'git remote' or
'git for-each-ref' for listing the branches for each remote and work on the
string list returned to get the same information.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>