cmd/mr_checkout: fix fetch reference existence check
When the tracking mode was enabled (`-t`) the reference being passed to
`git show-ref` was being wrongly composed, since the code was always
expecting a simple branch name instead of a full remote ref. This patch
fixes it by separating the check in two steps: 1) check if the local
branch to which the MR will be checked-out exists and 2) if tracking is
enabled, check if the remote ref already exists. Both are influenced by
`--force`.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>