root: remove git wrapping code
Remove the last piece of git passthrough code: in case a command or a
flag was not recognized it was sent to `git` for parsing. This patch
removes it and rely completely on code owned by lab. Git is still used
in some places for retrieving repo information and for cloning repos
with `lab clone` command, but no command is passed through to git
anymore. This change allow us to fully recommend users to run `lab` as
an Git alias instead.

At the same time, some testing code had to be changed to adhere to the
lack of passthrough code. Some were completely removed (were testing the
passthrough code) and others sligthly changed to use Git directly
instead of lab+passthrough.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>