util: print command flags enabled when on --debug
Default command flag values can be overriden by setting it on lab's
configuration file and since multiple configuration files are allowed
sometimes the mix of flags can be quite confusing. To help with it this
patch adds the print of all enabled flags when the user calls a command
with --debug enabled (the 'debug' flag itself is ignored though).

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>