cmd/mr_create: add --no-edit flag to not open editor
The default behavior when creating an MR is to open the editor regardless
the amount of commits:

- 1 commit: open editor with the commit's title and body already in place.
- 1+ commits: open editor with an empty message

This patch adds the --no-edit to allow users of the first option (single
commit) to tell lab to create the MR using the commit's message right away,
without opening the editor.

If --no-edit is used with the second option (multiple commits) lab exits
with a error message.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>