project_create_test: fix negative sign on testdata dir name
A random usigned interger is used for generating the testdata directory
name, but an int() casting is done right after and, because of that,
depending on the size of the generated uint, the int casting will make it a
negative number. Consequently, the negative sign is maintained in the
resulting string from the strconv.Itoa() conversion, leading to a directory
name with two dashes: "testdata--8339364144815363225".

Even though this name format isn't a problem to lab itself, it's used as
"project name" for the project_create command testing code, which isn't
supported by GitLab:

ERROR: project_create.go:94: POST https://gitlab.com/api/v4/projects: 400
{message: {path: [must not start or end with a special character and must
not contain consecutive special characters.]}}

This patch format the random uint directly to string without using the
intermediate int() casting.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>