mr list: Milestone exact match fix
The following error occurs with 'lab mr list' when looking for MRs with a
specific milestone:

[prarit@prarit rhel-8]$ lab mr list --state merged --milestone 'RHEL-8.7.0'
2022/04/09 08:58:45 ERROR: mr_list.go:120: Milestone 'RHEL-8.7.0' is ambiguous

There are two milestones with similar names in the project:

[prarit@prarit rhel-8]$ lab milestone list  | grep "8.7.0"
RHEL-8.7.0-test - Red Hat Enterprise Linux 8, release 8.7.0
RHEL-8.7.0 - Red Hat Enterprise Linux 8.7.0 release

The lab error occurs because the GitLab API List Milestones "Search"
option does not do exact matching and returns a list of possible matching
milestones.  Since no one else has hit this error it seems safe to look
for and return an exact match if the API returns multiple milestones.

Look for an exact match when searching for a milestone.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>