mr_show: fix base SHA on git-show due to merged MRs
For open merge-requests the "git-show" command ranging from the
<remote>/<target-branch> to the head MR commit is fine, but once the MR
is merged, the diff will be empty, since the code is present in
<remote>/<target-branch>. The user should be able to call `lab mr show
-p` and see the patches of that MR regardless it being merged or not,
like it's possible through the WebUI.

This patch fixes it by using the merge-request base commit SHA instead
of the target-branch. The MR base SHA points to the parent commit of the
first commit in the MR, meaning that it'll be always a valid SHA on the
target-branch. The head SHA will be kept the same: "mr.SHA".

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>