goreleaser: fix homebrew test case
On previous commits lab stopped to wrap git commands, thus the homebrew
test needed to be fixed in the point where a call to `lab ls-files` was
being made. At the same time, the whole test doesn't make sense
considering lab wasn't really being used. So, instead of only fixing the
`ls-files` portion, this patch is changing the test to only check the
lab version.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>