gitlab: remove redundant code on calculating perpage items
The GitLab API has a fixed ceil limit for the number of items in a
single response page; the only situation where we need to calculate a
specific number of items per page is when the user passes a non-multiple
number and greater than the ceil limit, meaning we need to shrink the
number to get exactly what the user requested. This patch fixes this
nitpick in the code and also reorder some condition check to avoid not
needed API requests.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>