cmd: improve testing to avoid environment test errors
Some tests are somewhat dependent on how the test repo was left after
the last test run, meaning that if something went wrong some tests might
fail just because the repo isn't in the state it was expected to.

This patch fix some of these cases and also improve some other tests to
retrieve only the required data instead of retrieving everything and
selecting the needed portion later.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>