util: Consider branch remote when guessing MR
When trying to guess the merge request that corresponds to a
local branch, we search for MRs that match the (upstream) branch
name.

But as branch names only have to be unique within a project, it
is possible that there are multiple merge requests from different
forks that use the same source branch name.

Luckily we do know which is the correct source project: It's the
one that corresponds to the remote that the local branch was pushed
to.